refactor `measures.py` into module for redability
